requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
sqlalchemy==2.0.26
pymysql==1.1.0
mysql-connector-python==8.2.0
protobuf==4.21.12
redis==5.0.1
python-jose==3.3.0
python-dotenv==1.0.0
httpx>=0.27.0,<0.29.0
passlib==1.7.4
bcrypt==4.0.1
pydantic==2.11.4
pydantic-settings==2.9.1
python-multipart==0.0.6
rich>=13.7.0
argparse>=1.4.0
requests>=2.32.0
asyncio>=3.4.3
langchain==0.3.25
langchain-core==0.3.59
langchain-community==0.3.24
langchain-ollama==0.3.2
ollama==0.4.8
google-generativeai>=0.7.0
tiktoken>=0.5.0
tenacity>=8.2.0
aiocache>=0.12.1